Added correct files in PhysicTools/PatAlgos/python/tools
